extract NestedAttributes inner classes to their own files, that file was getting long
